+++ src/main/java/com/nawabali/nawabali/repository/querydsl/comment/CommentDslRepositoryCustom.java
@@ -1,10 +1,10 @@
 package com.nawabali.nawabali.repository.querydsl.comment;
 
-import com.nawabali.nawabali.dto.querydsl.CommentDslDto;
+import com.nawabali.nawabali.dto.CommentDto;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
 public interface CommentDslRepositoryCustom {
     //findCommentByPostIdWithParentOrderByParentIdAscNullsFirstCreatedAtAsc
-    Slice<CommentDslDto.ResponseDto> findCommentsByPostId(Long postId, Pageable pageable);
+    Slice<CommentDto.GetResponseDto> findCommentsByPostId(Long postId, Pageable pageable);
 }
